SierraTransformerIntegrationTest
