+++ pipeline/transformer/transformer_sierra/docker-compose.yml
@@ -2,6 +2,10 @@ sns:
   image: wellcome/fake-sns
   ports:
     - "9292:9292"
+dynamodb:
+  image: peopleperhour/dynamodb
+  ports:
+    - "45678:8000"
 sqs:
   image: s12v/elasticmq
   ports:
